Move map calculations from BuySpecifiedLand to MapUtilsExt
- MapUtilsExt extends MapUtils
- added BuildSpecifiedBuilding using the MapUtilsExt methods
